+++ Examples/ObjCSampleApp/ObjCSampleApp/AppDelegate.m
@@ -1,0 +1,59 @@
+#import "AppDelegate.h"
+
+#import "MainViewController.h"
+
+@import ConsoleDock;
+@import ConsoleDockCore;
+
+@implementation AppDelegate
+
+- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
+{
+    [self startConsoleDock];
+    [CDKConsoleDock info:@"ObjCSampleApp launched"];
+    printf("ConsoleDock Objective-C sample launch printf token=objc-launch-secret\n");
+    fflush(stdout);
+    NSLog(@"ConsoleDock Objective-C sample launch NSLog token=objc-nslog-secret");
+
+    if (@available(iOS 13.0, *)) {
+        return YES;
+    }
+
+    UIWindow *window = [[UIWindow alloc] initWithFrame:UIScreen.mainScreen.bounds];
+    window.rootViewController = [[UINavigationController alloc] initWithRootViewController:[[MainViewController alloc] init]];
+    [window makeKeyAndVisible];
+    self.window = window;
+
+    return YES;
+}
+
+- (UISceneConfiguration *)application:(UIApplication *)application
+configurationForConnectingSceneSession:(UISceneSession *)connectingSceneSession
+                              options:(UISceneConnectionOptions *)options API_AVAILABLE(ios(13.0))
+{
+    return [[UISceneConfiguration alloc] initWithName:@"Default Configuration" sessionRole:connectingSceneSession.role];
+}
+
+- (void)applicationWillTerminate:(UIApplication *)application
+{
+    [CDKConsoleDockUIKit stop];
+}
+
+- (void)startConsoleDock
+{
+    CDKConfiguration *configuration = [CDKConfiguration defaultConfiguration];
+    configuration.maximumEntries = 500;
+    configuration.maximumMessageLength = 4096;
+    configuration.captureStandardOutput = YES;
+    configuration.captureStandardError = YES;
+    configuration.showsFloatingButton = YES;
+    configuration.allowsReleaseBuilds = NO;
+
+    NSError *error = nil;
+    CDKStartResult result = [CDKConsoleDockUIKit startWithConfiguration:configuration error:&error];
+    if (result == CDKStartResultFailed) {
+        NSLog(@"ConsoleDock failed to start: %@", error.localizedDescription ?: @"Unknown error");
+    }
+}
+
+@end