+++ Examples/ObjCSampleApp/ObjCSampleApp/SceneDelegate.h
@@ -1,0 +1,8 @@
+#import <UIKit/UIKit.h>
+
+API_AVAILABLE(ios(13.0))
+@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
+
+@property (nonatomic, strong) UIWindow *window;
+
+@end