+++ Examples/ObjCSampleApp/ObjCSampleApp/SceneDelegate.m
@@ -1,0 +1,21 @@
+#import "SceneDelegate.h"
+
+#import "MainViewController.h"
+
+@implementation SceneDelegate
+
+- (void)scene:(UIScene *)scene
+willConnectToSession:(UISceneSession *)session
+      options:(UISceneConnectionOptions *)connectionOptions
+{
+    if (![scene isKindOfClass:UIWindowScene.class]) {
+        return;
+    }
+
+    UIWindow *window = [[UIWindow alloc] initWithWindowScene:(UIWindowScene *)scene];
+    window.rootViewController = [[UINavigationController alloc] initWithRootViewController:[[MainViewController alloc] init]];
+    [window makeKeyAndVisible];
+    self.window = window;
+}
+
+@end